Fixed some conversion warnings; Added new warnings with compiler flag -Wextra :)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.3@4458 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/ANALYSIS/MAPMATCHING/PoseClusteringAffineSuperimposer.h
@@ -183,7 +183,7 @@ namespace OpenMS
 						++i
 					)
 			{
-				setProgress(10+Real(i)/model_map_size*100.f);
+				setProgress((UInt)(10+Real(i)/model_map_size*100.f));
 
 				// Adjust window in scene map
 				while ( k_low < scene_map_size &&
